Convert json BSD data for single individual to class individual

This function takes data from a json source and saves as a tibble with a person attribute.

convert_bds_ind(txt = "", format = "1.0", ...)

Arguments

txt
A JSON string, URL or file

format
String. JSON data schema version number. There are currently three schemas supported: 1.0, 1.1 and 2.0.

...
Passed down to jsonlite::fromJSON()

Value

A tibble with a person attribute.

Note

Deprecated. Use fetch_loc() instead.

Author

Stef van Buuren 2021

Examples

fn <- system.file("testdata", "client3.json", package = "james")
p <- convert_bds_ind(fn)
#> Warning: convert_bds_ind() is deprecated. Please use fetch_loc() instead.