Draw growth chart with individual data

The function draw_chart_ind() expect an input location from a previous call, and plots the individual data on the requested growth chart.

draw_chart_ind(loc = "", chartcode = "", curve_interpolation = TRUE, ...)

Arguments

loc

Alternative to txt. Location where input data is uploaded. Argument loc is deprecated and will disappear in Sept 2022; please use session instead.

chartcode

Optional. The code of the requested growth chart.

curve_interpolation

A logical indicating whether curve interpolation shoud be applied.

...

Ignored

Note

Deprecated. Please use the more comprehensive draw_chart() function.

See also