Request site containing personalised charts

Server-side function to construct a URL to a site that shows a personalised growth chart. The site includes a navigation bar so that the end user can interact to chart choices.

request_site(
  txt = "",
  loc = "",
  format = "1.0",
  upload = TRUE,
  host = NULL,
  ...
)

Arguments

txt
A JSON string, URL or file with the data in JSON format. The input data adhere to specification BDS JGZ 3.2.5, and are converted to JSON according to schema.

loc
Alternative to txt. Location where input data is uploaded and converted to internal server format.

format
String. JSON data schema version number. There are currently three schemas supported: 1.0, 1.1 and 2.0.

upload
Logical. If TRUE then request_site() will upload the data in txt and return a site address with the ?loc= query appended. Setting (FALSE) just appends ?txt= to the site url, thus deferring validation and conversion to internal representation to the site.

host
URL of the JAMES server. By default, host is the currently running server that processes the request.

...
Passed down to jsonlite::fromJSON()

Value

URL composed of JAMES server, possibly appended by query string starting with ?txt= or ?loc=.

Details

One of txt or loc needs to be specified. If both are given, txt takes precedence. If neither is given, then the function returns the base site without any data.

Note

This function does not yet work with Docker because we cannot get the host URL name. The output form ?txt= currently ignores the schema argument.

See also

Examples

fn <- system.file("testdata", "client3.json", package = "james")
js <- jsonlite::toJSON(jsonlite::fromJSON(fn), auto_unbox = TRUE)
url <- "https://groeidiagrammen.nl/ocpu/library/james/testdata/client3.json"
host <- "http://localhost"

# solutions that upload the data and create a URL with the `?loc=` query parameter
if (FALSE) {
# upload file - does not work (object 'bds_schema_str.json' not found)
site <- request_site(fn, host = host)
# browseURL(site)

# upload JSON string
site <- request_site(js, host = host)
site
# browseURL(site)

# upload URL
site <- request_site(url, host = host)
site
# browseURL(site)

# same, but in two steps, starting from file name
# this also works for js and url
resp <- jamesclient::upload_txt(fn, host = host)
loc <- jamesclient::get_url(resp, "location")
site <- request_site(loc = loc)
site
# browseURL(site)

# solutions that create an immediate ?txt=[..data..] query
# this method does not create a cache on the server
site <- request_site(fn, upload = FALSE)
site <- request_site(js, upload = FALSE)
site <- request_site(url, upload = FALSE)
# browseURL(site)
}