Screen growth curves according to JGZ guidelines

Screen growth curves according to JGZ guidelines

screen_curves(
  txt = "",
  loc = "",
  location = "",
  format = "1.0",
  legacy = TRUE,
  ...
)

Arguments

txt
A JSON string, URL or file with the data in JSON format. The input data adhere to specification BDS JGZ 3.2.5, and are converted to JSON according to schema.

loc
Alternative to txt. Location where input data is uploaded and converted to internal server format.

location
Legacy for loc

format
String. JSON data schema version number. There are currently three schemas supported: 1.0, 1.1 and 2.0.

legacy
Logical indicating whether legacy should be done.

...
Passed down to jsonlite::fromJSON()

Value

A JSON string containing a table with screening results

Note

Deprecated for consistency. Function returns JSON, whereas all other functions return the R object. The alternative screen_growth() requests only results from screening. The alternative custom_list() produces the same list as screen_curves, but does not convert the result to JSON.

Examples

# # example json
# fn <- system.file("testdata", "client3.json", package = "james")
# fn <- system.file("testdata", "Laura_S_dev.json", package = "james")
#
# # first upload, then screen
# r1 <- jamesclient::upload_txt(fn)
# location <- jamesclient::get_url(r1, "location")
# location
# screen_curves(loc = location)
#
# # upload & screen
# screen_curves(fn)