Uploads, parses, converts and stores data on the server

Uploads JSON data that adhere to the BDS-format, parses its contents, converts it to an internal JAMES target object, and stores the result on the server for further processing. The function is useful for caching input data over multiple requests to OpenCPU. The cached data feed into other JAMES functions by means of the "loc" argument. The server wipes the cached data after 2 hours.

upload_data(
  txt = "",
  auto_format = TRUE,
  format = "1.0",
  schema = NULL,
  append_ddi = FALSE,
  verbose = FALSE,
  ...
)

Arguments

txt

A JSON string, URL or file

auto_format

Logical. Should the format be read from the data? Default is TRUE.

format

String. JSON data schema version number. There are currently three schemas supported: "1.0", "1.1" and "2.0". Formats "1.0" and "1.1" are included for backward compatibility only. Use format = "2.0" for new applications.

schema

A file name (optionally including the path) with the JSON validation schema. The schema argument overrides format. The function extracts the version number for the basename, and overwrites the format argument by version number.

append_ddi

Should the DDI responses be appended?

verbose

Show verbose output for centile::y2z()

...

Used for additional parameters

Value

A tibble with an attribute called "person"

See also

Author

Stef van Buuren 2021

Examples

fn <- system.file("testdata", "client3.json", package = "james")
p <- upload_data(fn)