Derive advice on chart choice from data

The function loads individual data from an url, calculates the chartcode and returns a list of parsed chartcode and age range of the data. The function is called at initialization to automate setting of proper chart and analysis defaults according to the child data.

convert_tgt_chartadvice(
  txt = "",
  loc = "",
  format = "1.0",
  chartcode = "",
  selector = c("data", "chartcode"),
  ind_loc = "",
  ...
)

Arguments

txt

A JSON string, URL or file with the data in JSON format. The input data adhere to specification BDS JGZ 3.2.5, and are converted to JSON according to schema.

loc

Alternative to txt. Location where input data is uploaded and converted to internal server format.

format

String. JSON data schema version number. There are currently three schemas supported: 1.0, 1.1 and 2.0.

chartcode

Optional. The code of the requested growth chart.

selector

Either "chartcode", "data" or "derive". The function can calculate the chart code by looking at the child data (method "data") or user input (method "derive"). More in detail, the following behaviour decides between growth charts:

"data"

Calculate chart code from the individual data. This setting chooses the "optimal" chart for a given individual set of data.

"derive"

Calculate chart code from a combination of user data: chartgrp, agegrp, side, sex, etn, ga. The method does not use individual data. Use this setting when chart choice needs to be reactive on user input.

"chartcode"

Take string specified in chartcode

If there is a valid tgt object, then the function simply obeys the selector setting. If no valid tgt object is found, the "chartcode" argument is taken. However, if the "chartcode" is empty, then the function selects method "derive".

ind_loc

Legacy for loc. Use loc instead.

...

Passed down to jsonlite::fromJSON()

Value

A list with the following elements

population

A string identifying the population, e.g. 'NL','MA', 'TU' or 'PT'.

sex

A string "male", "female" or "undifferentiated".

design

A letter indicating the chart design: 'A' = 0-15m, 'B' = 0-4y, 'C' = 1-21y, 'D' = 0-21y, 'E' = 0-4ya.

side

A string indicating the side or yname: 'front', 'back', 'both', 'hgt', 'wgt', 'hdc', 'bmi', 'wfh'

language

The language in which the chart is drawn. Currently only "dutch" charts are implemented, but for population == "PT" we may also have "english".

week

A scalar indicating the gestational age at birth. Only used if population == "PT".

chartcode

A string indicating the chart code.

chartgrp

A string indicating the chart group, either "nl2010", "preterm" or "who".

agegrp

A string indicating the age group, either "0-15m", "0-4y", "1-21y" or "0-21y".

dnr

A string indicating the donor dataset for matching, either "smocc", "lollypop", "terneuzen" or "pops".

slider_list

A string indicating the set of slider labels, either "0_2", "0_4" or "0_29".

period

A character vector of two elements, indicating the first and last period for the matching analysis, e.g. like c("3m", "14m").

See also

Author

Stef van Buuren 2020